Restructure alert component
