+++ packages/react/src/index.ts
@@ -53,6 +53,7 @@ export * from "./tabs-list";
 export * from "./tabs-trigger";
 export * from "./text";
 export * from "./textarea";
+export * from "./toast";
 export * from "./toast-action";
 export * from "./toast-provider";
 export * from "./toast-title";
