+++ packages/web-components/package.json
@@ -18,51 +18,61 @@
     "./AccordionContent": "./dist/components/AccordionContent.js",
     "./AccordionItem": "./dist/components/AccordionItem.js",
     "./AccordionTrigger": "./dist/components/AccordionTrigger.js",
-    "./Alert": "./dist/Alert.js",
-    "./AlertDescription": "./dist/AlertDescription.js",
-    "./AlertDialog": "./dist/AlertDialog.js",
-    "./AlertTitle": "./dist/AlertTitle.js",
-    "./Avatar": "./dist/Avatar.js",
-    "./AnimatePresence": "./dist/AnimatePresence.js",
-    "./Badge": "./dist/Badge.js",
-    "./Box": "./dist/Box.js",
-    "./Button": "./dist/Button.js",
-    "./ButtonGroup": "./dist/ButtonGroup.js",
-    "./Checkbox": "./dist/Checkbox.js",
-    "./Chip": "./dist/Chip.js",
-    "./Code": "./dist/Code.js",
-    "./Dialog": "./dist/Dialog.js",
-    "./DialogBody": "./dist/DialogBody.js",
-    "./DialogFooter": "./dist/DialogFooter.js",
-    "./DialogTitle": "./dist/DialogTitle.js",
-    "./Field": "./dist/Field.js",
-    "./Flex": "./dist/Flex.js",
-    "./Grid": "./dist/Grid.js",
-    "./Heading": "./dist/Heading.js",
-    "./Indicator": "./dist/Indicator.js",
-    "./Input": "./dist/Input.js",
-    "./Kbd": "./dist/Kbd.js",
-    "./Link": "./dist/Link.js",
+    "./Alert": "./dist/components/Alert.js",
+    "./AlertTitle": "./dist/components/AlertTitle.js",
+    "./AlertDescription": "./dist/components/AlertDescription.js",
+    "./AlertDialog": "./dist/components/AlertDialog.js",
+    "./AnimatePresence": "./dist/components/AnimatePresence.js",
+    "./Avatar": "./dist/components/Avatar.js",
+    "./Badge": "./dist/components/Badge.js",
+    "./Box": "./dist/components/Box.js",
+    "./Button": "./dist/components/Button.js",
+    "./ButtonGroup": "./dist/components/ButtonGroup.js",
+    "./Checkbox": "./dist/components/Checkbox.js",
+    "./Chip": "./dist/components/Chip.js",
+    "./Code": "./dist/components/Code.js",
+    "./Dialog": "./dist/components/Dialog.js",
+    "./DialogBody": "./dist/components/DialogBody.js",
+    "./DialogFooter": "./dist/components/DialogFooter.js",
+    "./DialogTitle": "./dist/components/DialogTitle.js",
+    "./Field": "./dist/components/Field.js",
+    "./Flex": "./dist/components/Flex.js",
+    "./Grid": "./dist/components/Grid.js",
+    "./Heading": "./dist/components/Heading.js",
+    "./Indicator": "./dist/components/Indicator.js",
+    "./Input": "./dist/components/Input.js",
+    "./Kbd": "./dist/components/Kbd.js",
+    "./Link": "./dist/components/Link.js",
     "./Pagination": "./dist/components/Pagination.js",
-    "./Paper": "./dist/Paper.js",
+    "./Menu": "./dist/components/Menu.js",
+    "./MenuContent": "./dist/components/MenuContent.js",
+    "./MenuItem": "./dist/components/MenuItem.js",
+    "./MenuLabel": "./dist/components/MenuLabel.js",
+    "./MenuSeparator": "./dist/components/MenuSeparator.js",
+    "./MenuTrigger": "./dist/components/MenuTrigger.js",
+    "./Paper": "./dist/components/Paper.js",
     "./Popover": "./dist/components/Popover.js",
     "./PopoverContent": "./dist/components/PopoverContent.js",
     "./PopoverTrigger": "./dist/components/PopoverTrigger.js",
-    "./Progress": "./dist/Progress.js",
-    "./RadioGroup": "./dist/RadioGroup.js",
-    "./RadioGroupItem": "./dist/RadioGroupItem.js",
-    "./Search": "./dist/Search.js",
-    "./Separator": "./dist/Separator.js",
-    "./Skeleton": "./dist/Skeleton.js",
-    "./Switch": "./dist/Switch.js",
-    "./Tabs": "./dist/Tabs.js",
-    "./TabsContent": "./dist/TabsContent.js",
-    "./TabsList": "./dist/TabsList.js",
-    "./TabsTrigger": "./dist/TabsTrigger.js",
-    "./Textarea": "./dist/Textarea.js",
-    "./Text": "./dist/Text.js",
-    "./Tooltip": "./dist/Tooltip.js",
-    "./Transition": "./dist/Transition.js"
+    "./Progress": "./dist/components/Progress.js",
+    "./RadioGroup": "./dist/components/RadioGroup.js",
+    "./RadioGroupItem": "./dist/components/RadioGroupItem.js",
+    "./Search": "./dist/components/Search.js",
+    "./Separator": "./dist/components/Separator.js",
+    "./Skeleton": "./dist/components/Skeleton.js",
+    "./Switch": "./dist/components/Switch.js",
+    "./Tabs": "./dist/components/Tabs.js",
+    "./TabsContent": "./dist/components/TabsContent.js",
+    "./TabsList": "./dist/components/TabsList.js",
+    "./TabsTrigger": "./dist/components/TabsTrigger.js",
+    "./Text": "./dist/components/Text.js",
+    "./Textarea": "./dist/components/Textarea.js",
+    "./Toast": "./dist/components/Toast.js",
+    "./ToastAction": "./dist/components/ToastAction.js",
+    "./ToastProvider": "./dist/components/ToastProvider.js",
+    "./ToastTitle": "./dist/components/ToastTitle.js",
+    "./Tooltip": "./dist/components/Tooltip.js",
+    "./Transition": "./dist/components/Transition.js"
   },
   "types": "dist/index.d.ts",
   "scripts": {
