+++ packages/web-components/src/components/AlertDescription.ts
@@ -2,7 +2,7 @@ import { AlertDescription as AlertDescriptionComponent } from "@optiaxiom/react"
 
 import type { ComponentAttributes } from "./ComponentAttributes";
 
-import { register } from "./register";
+import { register } from "../register";
 
 export const AlertDescription = "ax-alert-description";
 register(AlertDescription, AlertDescriptionComponent);
