+++ packages/web-components/src/components/AlertTitle.ts
@@ -2,7 +2,7 @@ import { AlertTitle as AlertTitleComponent } from "@optiaxiom/react";
 
 import type { ComponentAttributes } from "./ComponentAttributes";
 
-import { register } from "./register";
+import { register } from "../register";
 
 export const AlertTitle = "ax-alert-title";
 register(AlertTitle, AlertTitleComponent);
